% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bintohex.R
\name{bintohex}
\alias{bintohex}
\title{this function convert a number binary 16bits a number hex}
\usage{
bintohex(binary)
}
\arguments{
\item{number_bin}{A string the number to convert}
}
\value{
A string the hexadecimal number
}
\description{
this function its works with a number binary and convert to hex representation.
}
\examples{

bintohex("10111010001")

#Returns "5D1"


}
