% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTransformation.R
\docType{data}
\name{Delay}
\alias{Delay}
\title{Delay is a local varibale for use the command filter}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10200 rows and 19 columns.
}
\usage{
Delay
}
\description{
Delay is a local varibale for use the command filter
}
\keyword{datasets}
