% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTransformation.R
\docType{data}
\name{flight_exercise}
\alias{flight_exercise}
\title{Solution 5.2.2 item 2:Compare air_time with arr_time - dep_time. What do you expect to see? What do you see? What do you need to do to fix it?}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 336776 rows and 3 columns.
}
\usage{
flight_exercise
}
\description{
Solution 5.2.2 item 2:Compare air_time with arr_time - dep_time. What do you expect to see? What do you see? What do you need to do to fix it?
}
\keyword{datasets}
