% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTransformation.R
\docType{data}
\name{fly_houston}
\alias{fly_houston}
\title{Solution 5.2.4 item 2:    Find all flights that flew to Houston (IAH or HOU)
fly_houston is local variable for use the command filter}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9313 rows and 19 columns.
}
\usage{
fly_houston
}
\description{
Solution 5.2.4 item 2:    Find all flights that flew to Houston (IAH or HOU)
fly_houston is local variable for use the command filter
}
\keyword{datasets}
