% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTransformation.R
\docType{data}
\name{Minipak}
\alias{Minipak}
\title{Creating a dataframe manual}
\format{
An object of class \code{data.frame} with 4 rows and 4 columns.
}
\usage{
Minipak
}
\description{
Creating a dataframe manual
}
\keyword{datasets}
