% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTransformation.R
\docType{data}
\name{UECCi}
\alias{UECCi}
\title{Creating a dataframe manual}
\format{
An object of class \code{data.frame} with 8 rows and 3 columns.
}
\usage{
UECCi
}
\description{
Creating a dataframe manual
}
\keyword{datasets}
