% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTransformation.R
\docType{data}
\name{vars}
\alias{vars}
\title{Solution 5.4.1 item 3:What does the any_of() function do? Why might it be helpful in conjunction with this vector?
 >[!NOTE]
the function any_of() select variables contained in a character vector. They are especially useful for programming with selecting functions}
\format{
An object of class \code{character} of length 5.
}
\usage{
vars
}
\description{
Solution 5.4.1 item 3:What does the any_of() function do? Why might it be helpful in conjunction with this vector?
 >[!NOTE]
the function any_of() select variables contained in a character vector. They are especially useful for programming with selecting functions
}
\keyword{datasets}
